refactored image downloading to store images by id. This is done using a hashmap
