+++ src/main/java/images/XMLParser.java
@@ -1,5 +1,7 @@
-package main;
+package images;
 
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
@@ -11,22 +13,23 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.IOException;
 import java.io.StringReader;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 public class XMLParser {
 
-    public List<String> parse(String XMLString){
+    public HashMap<String,String> parse(String XMLString){
 
         Document doc = convertToXML(XMLString);
-        List<String> urlList = new ArrayList<>();
+        HashMap<String,String> urlList = new HashMap<>();
         if (doc != null) {
             NodeList photos = doc.getElementsByTagName("photos");
             Element photosElement = (Element) photos.item(0);
             NodeList list = photosElement.getElementsByTagName("photo");
             for(int i=0;i<list.getLength();i++){
+                String id = list.item(i).getAttributes().getNamedItem("id").getTextContent();
                 String url = list.item(i).getAttributes().getNamedItem("url_m").getTextContent();
-                urlList.add(url);
+                urlList.put(id,url);
+
                 //System.out.println(url);
             }
             return urlList;
